GameManager에 있던 ui 관련 코드를 UiManager로 이동
