+++ src/com/brbr/brick/GameManager.java
@@ -26,14 +26,6 @@ public class GameManager {
     private InputManager inputManager;
     private AnimationManager animationManager;
 
-    private UILayer beforeLayer;
-    private UILayer pauseLayer;
-    private UILayer proceedingLayer;
-
-    private TextUI scoreUI;
-    private TextUI recordUI;
-    private ButtonUI pauseButton;
-
     public GameManager() {
         init();
     }
@@ -43,12 +35,12 @@ public class GameManager {
         frame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
         // TODO : init managers
         scene = new Scene();
+        uiManager = UIManager.getInstance(scene);
         renderer = new Renderer(scene);
         physicManager = new PhysicManager(scene);
         levelManager = new LevelManager(scene);
         animationManager = new AnimationManager(scene);
         inputManager = InputManager.getInstance();
-        uiManager = UIManager.getInstance();
 
         inputManager.setTarget(renderer);
         // init scene frame
@@ -59,111 +51,9 @@ public class GameManager {
 
         createDummyData();
 
-        setProceedingUI();
-        setBeforeUI();
-        setPauseUI();
-
-        beforeGame();
-
         frame.getContentPane().add(renderer);
         frame.setSize(GAME_WIDTH, GAME_HEIGHT);
         frame.setVisible(true);
-    }
-
-    private void setBeforeUI(){
-        beforeLayer = new UILayer();
-
-        Color backgroundColor = new Color(251, 136, 54);
-        Color textColor = Color.WHITE;
-
-        ButtonUI startButton = new ButtonUI("Game Start", new Vector2(GAME_WIDTH / 2 - 100, GAME_HEIGHT / 2 - 55),
-                20, 200, 50);
-        startButton.setBackgroundColor(backgroundColor);
-        startButton.setTextColor(textColor);
-
-        ButtonUI quitButton = new ButtonUI("Quit", new Vector2(GAME_WIDTH / 2 - 100, GAME_HEIGHT / 2),
-                20, 200, 50);
-        quitButton.setBackgroundColor(backgroundColor);
-        quitButton.setTextColor(textColor);
-
-        startButton.setButtonClickCallback(() -> {
-            startGame();
-        });
-
-        quitButton.setButtonClickCallback(() -> { });
-
-        beforeLayer.addButtonUI(startButton);
-        beforeLayer.addButtonUI(quitButton);
-
-        beforeLayer.background = true;
-
-        uiManager.addLayer(beforeLayer);
-    }
-
-    private void setProceedingUI(){
-        proceedingLayer = new UILayer();
-        recordUI = new TextUI("RECORD: " + scene.scoreManager.record,
-                new Vector2(scene.frameWidth / 2 - 48, 30), 20);
-        scoreUI = new TextUI("SCORE: 0", new Vector2(scene.frameWidth / 2 - 48, 60), 20);
-
-        ButtonUI pauseButton = new ButtonUI("Ⅱ", new Vector2(scene.frameWidth - 50, 15),
-                20, 40, 40);
-        pauseButton.setButtonClickCallback(() -> {
-            pauseGame();
-        });
-
-        proceedingLayer.addTextUI(recordUI);
-        proceedingLayer.addTextUI(scoreUI);
-        proceedingLayer.addButtonUI(pauseButton);
-
-        uiManager.addLayer(proceedingLayer);
-    }
-
-    private void setPauseUI(){
-        pauseLayer = new UILayer();
-
-        Color backgroundColor = new Color(251, 136, 54);
-        Color textColor = Color.WHITE;
-
-        ButtonUI resumeButton = new ButtonUI("resume", new Vector2(GAME_WIDTH / 2 - 90, GAME_HEIGHT / 2 - 40),
-                20, 180, 50);
-
-        resumeButton.setButtonClickCallback(() -> {
-            startGame();
-        });
-        resumeButton.setBackgroundColor(backgroundColor);
-        resumeButton.setTextColor(textColor);
-
-        pauseLayer.addButtonUI(resumeButton);
-
-        pauseLayer.background = true;
-
-        uiManager.addLayer(pauseLayer);
-    }
-
-    private void beforeGame(){
-        scene.gameStatus = Scene.BEFORE_GAME;
-        beforeLayer.setVisible(true);
-        pauseLayer.setVisible(false);
-        proceedingLayer.setClickUnable(true);
-    }
-
-    private void startGame(){
-        scene.gameStatus = Scene.PROCEEDING_GAME;
-        beforeLayer.setVisible(false);
-        pauseLayer.setVisible(false);
-        proceedingLayer.setClickUnable(false);
-    }
-
-    private void pauseGame(){
-        scene.gameStatus = Scene.PAUSE_GAME;
-        beforeLayer.setVisible(false);
-        pauseLayer.setVisible(true);
-        proceedingLayer.setClickUnable(true);
-    }
-
-    private void endGame(){
-        scene.gameStatus = Scene.END_GAME;
     }
 
     // dummy data TODO : remove
@@ -276,9 +166,6 @@ public class GameManager {
         // render
         renderer.repaint();
 
-        scoreUI.setText("SCORE: " + scene.scoreManager.score);
-        recordUI.setText("RECORD: " + scene.scoreManager.record);
-
         // TODO : 이동 논의
         for (GameObject gameObject : scene.gameObjectList) {
             if (gameObject instanceof Ball) {
@@ -293,6 +180,6 @@ public class GameManager {
         physicManager.handleInput(inputData);
     }
 
-    private final static int GAME_WIDTH = 605;
-    private final static int GAME_HEIGHT = 800;
+    public final static int GAME_WIDTH = 605;
+    public final static int GAME_HEIGHT = 800;
 }
